+++ lib/rpcs/v1_5_0.rs
@@ -11,7 +11,6 @@ pub(crate) mod speculative_exec;
 
 pub(crate) mod get_account {
     pub use crate::rpcs::v1_4_5::get_account::GetAccountResult;
-    pub(crate) use crate::rpcs::v1_4_5::get_account::{GetAccountParams, GET_ACCOUNT_METHOD};
 }
 
 pub(crate) mod get_auction_info {